Temporary MCGalaxy-protection-level fix
Made MCGalaxy-protection-level comboBox automatically select the option "Off". This is a temporary fix as doing this will not allow the owner to change this option in the PropertyWindow when Applying or Saving server properties. However this does fix the issue with server.properties failing to successfully save/update within the GUI.