Add bot property: Owner
When you add (or copy) a bot its owner is set to your name. You can no
longer edit the properties of bots that you do not own unless you are
the realm owner. An extra permission has been added in /bots to allow
editing bots that you do not own.
You may now only remove all bots at once if you are the realm owner or
have permission to edit bots that are not your own.
/WhoNick now allows checking the name of bots.
The bot's true name in /bots is shown as navy if the bot has a nickname.