Fix an incredibly extremely rare bug which caused zones to get lost on saving
The exporter works by counting number of physics checks with non-0 data, then writing that count to disc, and then writing the actual physics checks to disc. However, because the 'writing actual physics checks' reloaded the 'number of physics checks' instead of using the same 'number' as when it was counting number of physics checks with non-0 data, this could cause it to write more entries to disc than the count stated on the extremely rare chance that some other thread modified the checks list in-between writing the 'count' and the actual physics checks. This would then cause zones to not get loaded the next time the map was loaded from disc, because the importer would try interpreting that extra entry as a section identifier. Because that would not match any section identifiers it knew about, it would give up reading anymore data, and hence would not read the actual zones list section. Although I had taken steps to ensure physics could not change checks list while the map was being saved, I forgot that a user deleting a door block would still add a physics check to the list, which could thus extremely rarely trigger this bug. TL;DR multithreading is a pain.