Fix rare bug where doors would get stuck as air form on saving
This was caused by writing blocks to disc, then physics tick running in another thread, then writing physics checks to disc. This would cause physics checks to be written according to the blocks in the map after the tick finished, but the blocks in the file would be from before the tick executed.