/abort: Get rid of 'remove unsent pending block updates made by the player' code (Thanks Goodly)
Because the actual blocks in the level blocks arrays were still updated, all this resulted in was the appearance of the map in clients becoming out of sync with the actual blocks in the map. (i.e. the blocks were actually changed, but didn't appear changed to clients until /reload ing)